refactor(whatsrust): finish the Rust facade cleanup
